Update Makefile with sentinel pattern.
Dependencies in Make rely on lots of metadata from the filesystem to know
whether or not rebuilding a given target is necessary.
Sentinel files help by getting in between these checks and bottlenecking the
process in order to reduce the overall amount of checking required.
There is further reading in this article:
https://tech.davis-hansson.com/p/make/